Replaced coins() iterface for old sbtc pool

+++ contracts/Interfaces/curve/Curve.sol
@@ -45,7 +45,7 @@ interface ICurveFi {
         bool use_underlying
     ) external payable;
 
-    function coins(uint256) external view returns (address);
+    function coins(int128) external view returns (address);
     function pool() external view returns (address);
     function base_coins(uint256) external view returns (address);
     function underlying_coins(uint256) external view returns (address);
